# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "文档描述略",
        "title": "HelloC API (Swagger Example)",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9995",
    "basePath": "/api/v1",
    "paths": {
...
        "/videos/": {
            "get": {
                "description": "无权限",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "获取视频信息列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "个数",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"成功获取视频列表!\",\"results\":[{\"id\":1,\"nickname\":\"小明\",\"account\":\"xiaoming\",\"password\":\"123456\",\"email\":\"\",\"phone\":\"\",\"isAdmin\":false,\"photoFile\":\"\",\"createdTime\":\"\"}}]}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"参数错误/获取视频列表失败!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "普通用户可上传",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "上传视频信息",
                "parameters": [
                    {
                        "description": "desc",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddVideoParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"创建视频完成!\",\"results\":videoObject}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"message\": \"参数错误/创建失败!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/videos/{id}/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "获取视频信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"成功获取视频信息!\",\"results\":{\"id\":1,\"nickname\":\"小明\",\"account\":\"xiaoming\",\"password\":\"123456\",\"email\":\"\",\"phone\":\"\",\"isAdmin\":false,\"photoFile\":\"\",\"createdTime\":\"\"}}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"参数错误/视频不存在!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "只有管理员/或用户自己可以修改视频信息",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "修改视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ModifyVideoParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"成功修改用户!\", \"results\": videoObject}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"message\": \"参数错误\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "只有管理员或作者可以删除视频",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "删除视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"删除完成！\",\"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"message\": \"参数错误/删除失败!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/videos/{id}/check/": {
            "post": {
                "description": "管理员",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "审核视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否通过",
                        "name": "pass",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"视频审核完成\", \"results\":null}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"视频审核失败!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/videos/{id}/upload/": {
            "post": {
                "description": "管理员或用户",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "上传视频文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "视频文件",
                        "name": "videoFile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"成功上传视频文件!文件地址:url!\",\"results\":null}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"参数错误/视频不存在!/成功上传视频文件!但视频信息修改失败!\", \"results\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "v1.AddVideoParam": {
            "type": "object",
            "required": [
                "title",
                "userId"
            ],
            "properties": {
                "category": {
                    "description": "视频类型",
                    "type": "string"
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "duration": {
                    "description": "视频时长",
                    "type": "string"
                },
                "label": {
                    "description": "视频标签，多个自定义标签",
                    "type": "string"
                },
                "rm": {
                    "description": "视频格式(mp4,flv)",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "userId": {
                    "description": "视频上传用户",
                    "type": "integer"
                }
            }
        },
...
        "v1.ModifyVideoParam": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "category": {
                    "description": "视频类型",
                    "type": "string"
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "duration": {
                    "description": "视频时长",
                    "type": "string"
                },
                "label": {
                    "description": "视频标签，多个自定义标签",
                    "type": "string"
                },
                "rm": {
                    "description": "视频格式(mp4,flv)",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
...
    }
}